Decoupled Home page from page.js into its own page
This is setting up for a simple state-based solution to switch between the home page and the assistant interaction page.
